MWPW-134604: Avoids updating the original mdast node
The content node object that was passed in `addTrackChangesInfo(...)` to add the langstore/region track change properties was a reference to the object. So if that object existed in the mergedMdast already, it was updating the existing content node object with the track change info which it should not; thus creating the issue. The change makes sure that the track change is applied to the content node which is a new object created from an existing object pulled from the `hashToContentMap`. This new updated content node object is then pushed to `mergedMdast`.