MWPW-136621 | Send to Caas IMS error (Single page Send to Caas from sidekick)
Send To CaaS from sidekick triggers a postXDM call with user access token.
This token is validated by by the service, with prod as default.
However for hlx.page or hlx.live the access token is of stage (based on window.adobeid being currently set).
A user-ims-env header needs to be passed approiately for the service to validate against stage.
Change is to set the header as stage in case adobeid.environment is non prod.

Due to another jira MWPW-136658 tested this with localhost by updating below in libs/utils/sidekick.js
   const sendToCaasListener = async (e) => {
     const { host, project, ref: branch, repo, owner } = e.detail.data.config;
-    const { sendToCaaS } = await import('https://milo.adobe.com/tools/send-to-caas/send-to-caas.js');
+    const { sendToCaaS } = await import('../../tools/send-to-caas/send-to-caas.js');
     sendToCaaS({ host, project, branch, repo, owner }, loadScript, loadStyle);
   };

Resolves: MWPW-136621