MWPW-173110 Fix fragment loading fail in iframe
When loading a milo page into an iframe, sometimes fragments would fail to load and would instead display as text.

The reason for this is that `createTag` was using `instanceof HTMLElement`.  However the instanceof Element check fails because the Element constructor we're checking against belongs to a different JavaScript realm (outside iframe) than the one where the element was created. Using myEl.nodeType === Node.ELEMENT_NODE is a more reliable, cross-context way to determine if an object is an HTML element.